+++ packages/zowe-explorer/src/extension.ts
@@ -726,6 +726,19 @@ function initUSSProvider(context: vscode.ExtensionContext) {
         )
     );
     context.subscriptions.push(
+        vscode.commands.registerCommand("zowe.uss.pasteUssFile", async (node: IZoweUSSTreeNode) => {
+            ussActions.pasteUssFile(ussFileProvider, node);
+        })
+    );
+    context.subscriptions.push(
+        vscode.commands.registerCommand(
+            "zowe.uss.copyUssFile",
+            async (node: IZoweUSSTreeNode, nodeList: IZoweUSSTreeNode[]) => {
+                ussActions.copyUssFiles(node, nodeList, ussFileProvider);
+            }
+        )
+    );
+    context.subscriptions.push(
         vscode.workspace.onDidChangeConfiguration((e) => {
             ussFileProvider.onDidChangeConfiguration(e);
         })
@@ -893,7 +906,6 @@ function initSubscribers(context: vscode.ExtensionContext, theProvider: IZoweTre
         });
     }
 }
-
 /**
  * Called by VSCode on shutdown
  *