+++ packages/zowe-explorer/CHANGELOG.md
@@ -4,13 +4,19 @@ All notable changes to the "vscode-extension-for-zowe" extension will be documen
 
 ## TBD Release
 
--Multiple select for hide session [#1555](https://github.com/zowe/vscode-extension-for-zowe/issues/1555)
+### New features and enhancements
+
+### Bug fixes
+
+## `2.5.0`
 
 ### New features and enhancements
 
 - Added ability to filter jobs by status. Improved Job filtering User experience.[#1925](https://github.com/zowe/vscode-extension-for-zowe/issues/1925)
 - Added option to view PDS member attributes, and updated formatting for attributes webview. [#1577](https://github.com/zowe/vscode-extension-for-zowe/issues/1577)
 - Streamlined attribute viewing options into one feature - "Show Attributes".
+- Added multiple select copy/paste feature on uss view [#1549](https://github.com/zowe/vscode-extension-for-zowe/issues/1549)
+- Added multiple select for hide session [#1555](https://github.com/zowe/vscode-extension-for-zowe/issues/1555)
 
 ### Bug fixes